☀️ 8% – src: humanize fifth hundred-region batch

+++ src/080045d4.c
@@ -2,6 +2,7 @@
 
 s32 Func_080072f0(s32, s32, s32, s32);
 
-s32 Func_080045d4(s32 arg0, s32 arg1, s32 arg2) {
-    return Func_080072f0(arg0, arg1, arg2, 0x030001D8) << 8;
+s32 Func_080045d4(s32 arg0, s32 arg1, s32 arg2)
+{
+    return (s32)((u32)Func_080072f0(arg0, arg1, arg2, 0x030001D8) << 8);
 }
